contact-list: auto refresh when the modal is closed
